Make getJobParameters() abstract.
Previously, migrating from the old job system to the new job system
would result in JobParameters not being deserialized properly (even
though everything else was). So I've changed it so JobParameters are
upon request in an abstract method. This means we'll always be able to
get JobParameters, even from migrated jobs.

Tested the migration with several types of pending jobs, and all seems
well.