Save work results persistently if they are preempted.
JobScheduler can preempt our job and then reschedule the same task in
the future, even if preempted work ends up finishing. We were previously
handling this case when two jobs were running simultaneously, but we
failed to handle this if the jobs were run independently. To do so, we
need to store the results in persistent storage, since there may be
application restarts in between job attempts.